requirements.txt
flask
gunicorn
pyyaml
foursquare
facebook-sdk
elasticsearch
jinja2
requests
